Switching to library: complete

See merge request !1

+++ gadgetcli/init.go
@@ -6,24 +6,26 @@ import (
 	"gopkg.in/yaml.v2"
 	"io/ioutil"
 	"path/filepath"
+	"errors"
+	"github.com/nextthingco/libgadget"
 	log "github.com/sirupsen/logrus"
 )
 
 // Process the build arguments and execute build
-func GadgetInit(args []string, g *GadgetContext) error {
+func GadgetInit(args []string, g *libgadget.GadgetContext) error {
 
 	initUu1 := uuid.NewV4()
 	initUu2 := uuid.NewV4()
 
 
-	log.Info("[INIT ]  Creating new project:")
+	log.Info("Creating new project:")
 
 	g.WorkingDirectory, _ = filepath.Abs(g.WorkingDirectory)
 	initName := filepath.Base(g.WorkingDirectory)
 	
-	log.Infof("[INIT ]    in %s", g.WorkingDirectory)
+	log.Infof("  in %s", g.WorkingDirectory)
 
-	initConfig := TemplateConfig(initName, fmt.Sprintf("%s", initUu1), fmt.Sprintf("%s", initUu2))
+	initConfig := libgadget.TemplateConfig(initName, fmt.Sprintf("%s", initUu1), fmt.Sprintf("%s", initUu2))
 
 	outBytes, err := yaml.Marshal(initConfig)
 	if err != nil {
@@ -31,6 +33,24 @@ func GadgetInit(args []string, g *GadgetContext) error {
 	}
 	
 	fileLocation := fmt.Sprintf("%s/gadget.yml", g.WorkingDirectory)
+	
+	gadgetFileExists, err := libgadget.PathExists(fileLocation)
+	
+	if gadgetFileExists {
+			
+		log.WithFields(log.Fields{
+			"function": "GadgetInit",
+			"location": fileLocation,
+			"init-stage": "overwriting file",
+		}).Debug("gadget.yml already exists in this location")
+
+
+		log.Errorf("There's already a config file here [%s]", fileLocation)
+		log.Warnf("Remove %s if you'd like to init again", fileLocation)
+		
+		err = errors.New("Tried to overwrite pre-existing configuration file")
+		return err		
+	}
 	
 	err = ioutil.WriteFile(fileLocation, outBytes, 0644)
 	if err != nil {
